Updated index.jade
Clarified the "code blocks" introduction to what actually needs to happen. At a glance, the tab/spaces aren't obvious to what's the difference between the markdown and the final rendering. It took me looking up another resource to understand what was done. :)

The other examples don't have a problem given they rely on the introduction to visible characters.